update odoo 16 , 20221228

+++ source/odoo/addons/mail/static/src/models/chatter.js
@@ -71,26 +71,11 @@ registerModel({
          *
          * @param {MouseEvent} ev
          */
-        onClickScheduleActivity(ev) {
-            const action = {
-                type: 'ir.actions.act_window',
-                name: this.env._t("Schedule Activity"),
-                res_model: 'mail.activity',
-                view_mode: 'form',
-                views: [[false, 'form']],
-                target: 'new',
-                context: {
-                    default_res_id: this.thread.id,
-                    default_res_model: this.thread.model,
-                },
-                res_id: false,
-            };
-            return this.env.services.action.doAction(
-                action,
-                {
-                    onClose: () => this.reloadParentView(),
-                }
-            );
+        async onClickScheduleActivity(ev) {
+            await this.messaging.openActivityForm({ thread: this.thread });
+            if (this.exists()) {
+                this.reloadParentView();
+            }
         },
         /**
          * Handles click on "send message" button.
@@ -187,6 +172,9 @@ registerModel({
          * @private
          */
         _onThreadIdOrThreadModelChanged() {
+            if (!this.threadModel) {
+                return;
+            }
             if (this.threadId) {
                 if (this.thread && this.thread.isTemporary) {
                     this.thread.delete();
@@ -276,6 +264,9 @@ registerModel({
         }),
         dropZoneView: one('DropZoneView', {
             compute() {
+                if (!this.thread) {
+                    return clear();
+                }
                 if (this.useDragVisibleDropZone.isVisible) {
                     return {};
                 }
@@ -396,6 +387,15 @@ registerModel({
         }),
         scrollPanelRef: attr(),
         /**
+         * Determines whether the view should reload after file changed in this chatter,
+         * such as from a file upload.
+         */
+        shouldReloadParentFromFileChanged: attr({
+            compute() {
+                return this.hasParentReloadOnAttachmentsChanged;
+            },
+        }),
+        /**
          * Determines the `Thread` that should be displayed by `this`.
          */
         thread: one('Thread'),
@@ -418,6 +418,9 @@ registerModel({
          */
         threadViewer: one('ThreadViewer', {
             compute() {
+                if (!this.thread) {
+                    return clear();
+                }
                 return {
                     hasThreadView: this.hasThreadView,
                     order: 'desc',
@@ -425,10 +428,11 @@ registerModel({
                 };
             },
             inverse: 'chatter',
-            required: true,
         }),
         topbar: one('ChatterTopbar', {
-            default: {},
+            compute() {
+                return this.thread ? {} : clear();
+            },
             inverse: 'chatter',
         }),
         useDragVisibleDropZone: one('UseDragVisibleDropZone', {
