+++ source/odoo/addons/stock/models/stock_move_line.py
@@ -368,7 +368,9 @@ class StockMoveLine(models.Model):
                     # TODO: make package levels less of a pain and fix this
                     package_level = ml.package_level_id
                     ml.package_level_id = False
-                    package_level.unlink()
+                    # Only need to unlink the package level if it's empty. Otherwise will unlink it to still valid move lines.
+                    if not package_level.move_line_ids:
+                        package_level.unlink()
 
         # When we try to write on a reserved move line any fields from `triggers` or directly
         # `reserved_uom_qty` (the actual reserved quantity), we need to make sure the associated
@@ -457,18 +459,12 @@ class StockMoveLine(models.Model):
 
         res = super(StockMoveLine, self).write(vals)
 
-        # Update scrap object linked to move_lines to the new quantity.
-        if 'qty_done' in vals:
-            for move in self.mapped('move_id'):
-                if move.scrapped:
-                    move.scrap_ids.write({'scrap_qty': move.quantity_done})
-
         # As stock_account values according to a move's `product_uom_qty`, we consider that any
         # done stock move should have its `quantity_done` equals to its `product_uom_qty`, and
         # this is what move's `action_done` will do. So, we replicate the behavior here.
         if updates or 'qty_done' in vals:
             moves = self.filtered(lambda ml: ml.move_id.state == 'done').mapped('move_id')
-            moves |= self.filtered(lambda ml: ml.move_id.state not in ('done', 'cancel') and ml.move_id.picking_id.immediate_transfer and not ml.reserved_uom_qty).mapped('move_id')
+            moves |= self.filtered(lambda ml: ml.move_id.state not in ('done', 'cancel') and ml.move_id.picking_id.immediate_transfer).mapped('move_id')
             for move in moves:
                 move.product_uom_qty = move.quantity_done
             next_moves._do_unreserve()
@@ -619,12 +615,12 @@ class StockMoveLine(models.Model):
             lines |= picking_id.move_line_ids.filtered(lambda ml: ml.product_id == self.product_id and (ml.lot_id or ml.lot_name))
         return lines
 
-    def _prepare_new_lot_vals(self):
+    def _get_value_production_lot(self):
         self.ensure_one()
         return {
-            'name': self.lot_name,
-            'product_id': self.product_id.id,
             'company_id': self.company_id.id,
+            'name': self.lot_name,
+            'product_id': self.product_id.id
         }
 
     def _create_and_assign_production_lot(self):
@@ -639,8 +635,7 @@ class StockMoveLine(models.Model):
             key_to_mls[key] |= ml
             if ml.tracking != 'lot' or key not in key_to_index:
                 key_to_index[key] = len(lot_vals)
-                lot_vals.append(ml._prepare_new_lot_vals())
-
+                lot_vals.append(ml._get_value_production_lot())
         lots = self.env['stock.lot'].create(lot_vals)
         for key, mls in key_to_mls.items():
             lot = lots[key_to_index[key]].with_prefetch(lots._ids)   # With prefetch to reconstruct the ones broke by accessing by index
